Simplify Skip Link
