+++ src/components/skip-link/_skip-link.scss
@@ -1,6 +1,5 @@
-.sg-skip-nav {
+.sg-skip-link {
   user-select: 'none';
-
   position: absolute;
   clip: rect(0 0 0 0);
   border: 0;
