--- src/components/skip-nav/SkipNav.stories.mdx
@@ -1,44 +1,0 @@
-import {ArgsTable, Meta, Story, Canvas} from '@storybook/addon-docs';
-
-import SkipNav from './SkipNav';
-
-<Meta title="Components/SkipNav" component={SkipNav} />
-
-# Skiplink
-
-- [Accessibility](#accessibility)
-
-## Overview
-
-<Canvas>
-  <Story name="single link">
-    {args => {
-      return (
-        <div>
-          <h2>Click tab to focus on skiplink thus make it visible</h2>
-          <SkipNav links={{name: 'skip to main content', href: '#main'}} />
-          <div id="main" tabindex="-1" />
-        </div>
-      );
-    }}
-  </Story>
-  <Story name="multiple links">
-    {args => {
-      return (
-        <div>
-          <h2>Click tab to focus on skiplink and make it visible</h2>
-          <SkipNav
-            links={[
-              {name: 'skip to main content', href: '#main'},
-              {name: 'skip to secondary content', href: '#secondary'},
-              {name: 'skip to tertiary content', href: '#tertiary'},
-            ]}
-          />
-          <div id="main" tabindex="-1" />
-          <div id="secondary" tabindex="-1" />
-          <div id="tertiary" tabindex="-1" />
-        </div>
-      );
-    }}
-  </Story>
-</Canvas>
