+++ src/sass/main.scss
@@ -62,4 +62,4 @@ $sgFontsPath: 'fonts/' !default;
 @import '../components/dialog/dialog';
 @import '../components/dialog/dialog-no-scroll';
 @import '../components/transition/transition';
-@import '../components/skip-nav/skip-nav';
+@import '../components/skip-link/skip-link';
